[BF] Add network connection change listener

+++ data/src/main/java/com/gmail/maystruks08/data/repository/SettingsRepositoryImpl.kt
@@ -7,6 +7,7 @@ import com.gmail.maystruks08.data.local.dao.CheckpointDAO
 import com.gmail.maystruks08.data.mappers.toCheckpointTable
 import com.gmail.maystruks08.data.remote.FirestoreApi
 import com.gmail.maystruks08.data.toDataClass
+import com.gmail.maystruks08.domain.NetworkUtil
 import com.gmail.maystruks08.domain.entities.Checkpoint
 import com.gmail.maystruks08.domain.entities.CheckpointType
 import com.gmail.maystruks08.domain.entities.ResultOfTask
@@ -21,36 +22,40 @@ class SettingsRepositoryImpl @Inject constructor(
     private val preferences: ConfigPreferences,
     private val settingsCache: SettingsCache,
     private val firebaseAuth: FirebaseAuth,
-    private val checkpointDAO: CheckpointDAO
+    private val checkpointDAO: CheckpointDAO,
+    private val networkUtil: NetworkUtil
 ) : SettingsRepository {
 
     override suspend fun updateConfig(): ResultOfTask<Exception, SettingsRepository.Config> {
         return ResultOfTask.build {
-            val checkpointsSnapshot = awaitTaskResult(firestoreApi.getCheckpoints())
-            checkpointsSnapshot.data?.toDataClass<HashMap<String, Checkpoint>?>()?.let { hashMap ->
-                val checkpoints = hashMap.values
-                checkpointDAO.deleteCheckpoints()
-                checkpointDAO.insertAllOrReplace(checkpoints.map { it.toCheckpointTable() })
-                settingsCache.checkpoints = checkpoints.filter { it.type == CheckpointType.NORMAL }.sortedBy { it.id }
-                settingsCache.checkpointsIronPeople = checkpoints.filter { it.type == CheckpointType.IRON }.sortedBy { it.id }
-            }
-            firebaseAuth.currentUser?.uid?.let { currentUserId ->
-                val result = awaitTaskResult(firestoreApi.getCheckpointsSettings(currentUserId))
-                val startDateResult = awaitTaskResult(firestoreApi.getDateOfStart())
+            if(networkUtil.isOnline()) {
+                val checkpointsSnapshot = awaitTaskResult(firestoreApi.getCheckpoints())
+                checkpointsSnapshot.data?.toDataClass<HashMap<String, Checkpoint>?>()?.let { hashMap ->
+                        val checkpoints = hashMap.values
+                        checkpointDAO.deleteCheckpoints()
+                        checkpointDAO.insertAllOrReplace(checkpoints.map { it.toCheckpointTable() })
+                        settingsCache.checkpoints =
+                            checkpoints.filter { it.type == CheckpointType.NORMAL }.sortedBy { it.id }
+                        settingsCache.checkpointsIronPeople = checkpoints.filter { it.type == CheckpointType.IRON }.sortedBy { it.id }
+                }
+                firebaseAuth.currentUser?.uid?.let { currentUserId ->
+                    val result = awaitTaskResult(firestoreApi.getCheckpointsSettings(currentUserId))
+                    val startDateResult = awaitTaskResult(firestoreApi.getDateOfStart())
 
-                val checkpointId = result["checkpointId"] as? HashMap<*, *>
-                val checkpointIronPeopleId = result["checkpointIronPeopleId"] as? HashMap<*, *>
-                val startDate = (startDateResult["Start at"] as? com.google.firebase.Timestamp)?.toDate()
+                    val checkpointId = result["checkpointId"] as? HashMap<*, *>
+                    val checkpointIronPeopleId = result["checkpointIronPeopleId"] as? HashMap<*, *>
+                    val startDate = (startDateResult["Start at"] as? com.google.firebase.Timestamp)?.toDate()
 
-                if(checkpointId != null) preferences.saveCurrentCheckpointId((checkpointId.values.first() as Long).toInt())
-                if(checkpointIronPeopleId != null)  preferences.saveCurrentIronPeopleCheckpointId((checkpointIronPeopleId.values.first() as Long).toInt())
-                if(startDate != null) preferences.saveDateOfStart(startDate.time)
+                    if (checkpointId != null) preferences.saveCurrentCheckpointId((checkpointId.values.first() as Long).toInt())
+                    if (checkpointIronPeopleId != null) preferences.saveCurrentIronPeopleCheckpointId((checkpointIronPeopleId.values.first() as Long).toInt())
+                    if (startDate != null) preferences.saveDateOfStart(startDate.time)
+                }
             }
-            getConfig()
+            getCachedConfig()
         }
     }
 
-    override suspend fun getConfig(): SettingsRepository.Config {
+    override suspend fun getCachedConfig(): SettingsRepository.Config {
         val currentCheckpointId = preferences.getCurrentCheckpoint()
         val currentIronPeopleCheckpointId = preferences.getCurrentIronPeopleCheckpoint()
         val startDate = Date(preferences.getDateOfStart())
@@ -68,11 +73,16 @@ class SettingsRepositoryImpl @Inject constructor(
     override suspend fun changeCurrentCheckpoint(checkpointNumber: Int) {
         settingsCache.currentCheckpoint = settingsCache.checkpoints[checkpointNumber]
         preferences.saveCurrentCheckpointId(checkpointNumber)
-        firebaseAuth.currentUser?.uid?.let { currentUserId ->
-            try {
-                firestoreApi.saveCheckpointsSettings(currentUserId, SettingsRepository.Config(checkpointNumber, null, null))
-            } catch (e: Exception) {
-                Timber.e(e)
+        if(networkUtil.isOnline()){
+            firebaseAuth.currentUser?.uid?.let { currentUserId ->
+                try {
+                    firestoreApi.saveCheckpointsSettings(
+                        currentUserId,
+                        SettingsRepository.Config(checkpointNumber, null, null)
+                    )
+                } catch (e: Exception) {
+                    Timber.e(e)
+                }
             }
         }
     }
@@ -80,11 +90,16 @@ class SettingsRepositoryImpl @Inject constructor(
     override suspend fun changeCurrentCheckpointForIronPeoples(checkpointNumber: Int) {
         settingsCache.currentIronPeopleCheckpoint = settingsCache.checkpointsIronPeople[checkpointNumber]
         preferences.saveCurrentIronPeopleCheckpointId(checkpointNumber)
-        firebaseAuth.currentUser?.uid?.let { currentUserId ->
-            try {
-                firestoreApi.saveCheckpointsSettings(currentUserId, SettingsRepository.Config(null, checkpointNumber, null))
-            } catch (e: Exception) {
-                Timber.e(e)
+        if(networkUtil.isOnline()) {
+            firebaseAuth.currentUser?.uid?.let { currentUserId ->
+                try {
+                    firestoreApi.saveCheckpointsSettings(
+                        currentUserId,
+                        SettingsRepository.Config(null, checkpointNumber, null)
+                    )
+                } catch (e: Exception) {
+                    Timber.e(e)
+                }
             }
         }
     }
@@ -92,10 +107,12 @@ class SettingsRepositoryImpl @Inject constructor(
     override suspend fun changeStartDate(date: Date) {
         settingsCache.dateOfStart = date
         preferences.saveDateOfStart(date.time)
-        try {
-            firestoreApi.saveDateOfStart(date)
-        } catch (e: Exception) {
-            Timber.e(e)
+        if(networkUtil.isOnline()) {
+            try {
+                firestoreApi.saveDateOfStart(date)
+            } catch (e: Exception) {
+                Timber.e(e)
+            }
         }
     }
 }
